# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Light Speed Ghost API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /writing/generate:
    post:
      operationId: generatePaper
      tags: [writing]
      summary: Generate a new academic paper with citations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePaperBody"
      responses:
        "200":
          description: Generated paper content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedPaper"
  /writing/outline:
    post:
      operationId: generateOutline
      tags: [writing]
      summary: Generate a paper outline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOutlineBody"
      responses:
        "200":
          description: Generated outline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedOutline"
...
components:
  schemas:
...
    GeneratePaperBody:
      type: object
      properties:
        topic:
          type: string
        subject:
          type: string
        paperType:
          type: string
          enum: [research, essay, thesis, literature_review, report]
        length:
          type: string
          enum: [short, medium, long]
        citationStyle:
          type: string
          enum: [apa, mla, chicago, harvard, ieee]
        additionalInstructions:
          type: string
        isStem:
          type: boolean
      required: [topic, subject, paperType, length, citationStyle]
    GeneratedPaper:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        bibliography:
          type: string
        wordCount:
          type: integer
        documentId:
          type: integer
      required: [title, content, citations, bibliography, wordCount]
    Citation:
      type: object
      properties:
        id:
          type: string
        authors:
          type: string
        title:
          type: string
        year:
          type: integer
        source:
          type: string
        url:
          type: string
        formatted:
          type: string
      required: [id, authors, title, year, source, formatted]
    GenerateOutlineBody:
      type: object
      properties:
        topic:
          type: string
        subject:
          type: string
        paperType:
          type: string
          enum: [research, essay, thesis, literature_review, report]
      required: [topic, subject, paperType]
    GeneratedOutline:
      type: object
      properties:
        title:
          type: string
        sections:
          type: array
          items:
            type: object
            properties:
              heading:
                type: string
              subsections:
                type: array
                items:
                  type: string
            required: [heading]
      required: [title, sections]
...