# lib/api-zod/src/generated/types/citation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Light Speed Ghost API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Citation {
  id: string;
  authors: string;
  title: string;
  year: number;
  source: string;
  url?: string;
  formatted: string;
}