# lib/api-zod/src/generated/types/generatePaperBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Light Speed Ghost API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratePaperBodyCitationStyle } from "./generatePaperBodyCitationStyle";
import type { GeneratePaperBodyLength } from "./generatePaperBodyLength";
import type { GeneratePaperBodyPaperType } from "./generatePaperBodyPaperType";

export interface GeneratePaperBody {
  topic: string;
  subject: string;
  paperType: GeneratePaperBodyPaperType;
  length: GeneratePaperBodyLength;
  citationStyle: GeneratePaperBodyCitationStyle;
  additionalInstructions?: string;
  isStem?: boolean;
}